commands refactor, fixes in Character Load
