+++ sql/DATA_FATE_ACCEL_CHARS.SQL
@@ -1,10 +1,9 @@
 DROP TABLE DATA_FATE_ACCEL_CHARS;
 
 CREATE TABLE DATA_FATE_ACCEL_CHARS (
-        ID int NOT NULL AUTO_INCREMENT
-        , SERVER VARCHAR(200)
+        SERVER VARCHAR(200)
         , CHARACTER_NAME VARCHAR(200)
-        , CHARACTER_SHORT_NAME VARCHAR(200) UNIQUE
+        , CHARACTER_SHORT_NAME VARCHAR(200) 
         , DEFAULT_DIE_CODE VARCHAR(20)
         , ASPECT_1 VARCHAR(200)
         , ASPECT_2 VARCHAR(200)
@@ -22,5 +21,5 @@ CREATE TABLE DATA_FATE_ACCEL_CHARS (
         , CONSEQUENCE_2 VARCHAR(200)
         , CONSEQUENCE_4 VARCHAR(200)
         , CONSEQUENCE_6 VARCHAR(200)
-        , PRIMARY KEY (ID)
+        , PRIMARY KEY (SERVER, CHARACTER_SHORT_NAME)
 );
